laba3_1/requirements.txt
Django
djangorestframework
pyjwt